downloader/requirements.txt
certifi==2025.6.15
charset-normalizer==3.4.2
idna==3.10
PyYAML==6.0.2
requests==2.32.4
urllib3==2.5.0
certifi==2025.6.15
charset-normalizer==3.4.2
idna==3.10
PyYAML==6.0.2
requests==2.32.4
urllib3==2.5.0
